my_solution/requirements.txt
camelot_py==0.9.0
numpy==1.26.2
transformers==4.35.2
